Moved ndt_typ_from_pylist out of the type_conversions files. Added a comment block about the purposes of the conversion files.
